Update ksysrqd.c
Some patch for Kernel 5.4; however, it seems that kernel_accept does not interrupt when getting a signal. Therefore the module may not unload before you don't open a telnet connection, at least to release the loop. It may not matter for production; it is just bugging in development, and I am still looking for a long-time fix here.